+++ common/src/KokkosFFT_utils.hpp
@@ -147,22 +147,6 @@ bool are_valid_axes(const ViewType& /*view*/,
   return is_valid;
 }
 
-/// \brief Check if transpose is needed or not
-/// If a map is contiguous and in ascending order (e.g. {0, 1, 2}),
-/// we do not need transpose
-/// \tparam IndexType The integer type used for map
-/// \tparam DIM The dimensionality of the axes
-///
-/// \param[in] map The map used for permutation
-template <typename IndexType, std::size_t DIM>
-bool is_transpose_needed(const std::array<IndexType, DIM>& map) {
-  static_assert(std::is_integral_v<IndexType>,
-                "is_transpose_needed: IndexType must be an integral type.");
-  std::array<IndexType, DIM> contiguous_map;
-  std::iota(contiguous_map.begin(), contiguous_map.end(), 0);
-  return map != contiguous_map;
-}
-
 template <typename ContainerType, typename ValueType>
 std::size_t get_index(const ContainerType& values, const ValueType value) {
   using value_type = KokkosFFT::Impl::base_container_value_type<ContainerType>;
